Extinctions analysis for trophic networks

The SimulateExtinctions function, can be used to test how the order of species extinctions might affect the stability of the network by comparing The extintion history and checking for secondary extinctions.

SimulateExtinctions(
  Network,
  Method,
  Order = NULL,
  clust.method = "cluster_infomap"
)

Arguments

Network
a network representation as a an adyacency matrix, edgelist, or a network object

Method
a character with the options Mostconnected, Oredered, or Random

Order
this should be NULL, unless using the Ordered method, in that case it should be a vector with the order of extinctions by ID

clust.method
a character with the options cluster_edge_betweenness, cluster_spinglass, cluster_label_prop or cluster_infomap, defaults to cluster_infomap

Value

exports data frame with the characteristics of the network after every extintion. The resulting data frame contains 11 columns that incorporate the topological index, the secondary extinctions, predation release, and total extinctions of the network in each primary extinction.

Details

When method is Mostconnected, it takes a network and it calculates wich node is the most connected of the network, using total degree. Then remove the most connected node, and calculates the the topological indexes of the network and the number of secundary extintions (how many species have indegree 0, without considered primary producers). After that, remove the nodes that were secondarily extinct in the previous step and recalculate which is the new most connected node and so on, until the number of links in the network is zero.

When method is Ordered, it takes a network, and extinguishes nodes using a custom order, then it calculates the secondary extinctions and plots the accumulated secondary extinctions.

When clust.method = cluster_edge_betweenness computes the network modularity using cluster_edge_betweenness methods from igraph to detect communities When clust.method = cluster_spinglass computes the network modularity using cluster_spinglass methods from igraph to detect communities, here the number of spins are equal to the nerwork size When clust.method = cluster_label_prop computes the network modularity using cluster_label_prop methods from igraph to detect communities When clust.method = cluster_infomap computes the network modularity using cluster_infomap methods from igraph to detect communities, here the number of nb.trials are equal to the nerwork size

Author

Derek Corcoran <derek.corcoran.barrios@gmail.com>

M. Isidora Ávila-Thieme <msavila@uc.cl>

Examples

# Mostconnected example
data("net")
SimulateExtinctions(Network = net, Method = "Mostconnected", clust.method = "cluster_infomap")

#> [1] 1
#> [1] 2
#> [1] 3
#> [1] 4

#>   Spp S L          C Link_density SecExt Pred_release Iso_nodes Modularity
#> 1   6 9 7 0.08641975    0.7777778      1            0         1          0
#> 2   7 7 4 0.08163265    0.5714286      0            2         2          0
#> 3   5 6 2 0.05555556    0.3333333      1            3         3          0
#> 4   2 4 0 0.00000000    0.0000000      1            3         4          0
#>   AccSecExt NumExt TotalExt
#> 1         1      1        2
#> 2         1      2        3
#> 3         2      3        5
#> 4         3      4        7

#first Ordered example
data("net")
SimulateExtinctions(Network = net, Order = c(1,2,3,4,5,6,7,8,9,10), Method = "Ordered" , clust.method = "cluster_infomap")

#> Warning: the condition has length > 1 and only the first element will be used

#> 1

#> Warning: the condition has length > 1 and only the first element will be used

#> 2

#> Warning: the condition has length > 1 and only the first element will be used

#> 3

#> Warning: the condition has length > 1 and only the first element will be used

#> 4

#> Warning: the condition has length > 1 and only the first element will be used

#> 5

#>   Spp S L         C Link_density SecExt Pred_release Iso_nodes Modularity
#> 1   1 9 9 0.1111111    1.0000000      1            0         0          0
#> 2   2 7 6 0.1224490    0.8571429      1            0         0          0
#> 3   3 5 3 0.1200000    0.6000000      2            0         1          0
#> 4   4 2 1 0.2500000    0.5000000      1            0         0          0
#> 5   5 1 0 0.0000000    0.0000000      1            0         1          0
#>   AccSecExt NumExt TotalExt
#> 1         1      1        2
#> 2         2      2        4
#> 3         4      3        7
#> 4         5      4        9
#> 5         6      5       11

#Second Ordered example
data("net")
SimulateExtinctions(Network = net, Order = c(2,8,9), Method = "Ordered", clust.method = "cluster_infomap")

#> Warning: the condition has length > 1 and only the first element will be used

#> 1

#> Warning: the condition has length > 1 and only the first element will be used

#> 2

#> Warning: the condition has length > 1 and only the first element will be used

#> 3

#>   Spp S L          C Link_density SecExt Pred_release Iso_nodes Modularity
#> 1   2 9 8 0.09876543    0.8888889      1            0         0      0.000
#> 2   8 7 4 0.08163265    0.5714286      1            1         1      0.375
#> 3   9 6 4 0.11111111    0.6666667      0            1         0      0.375
#>   AccSecExt NumExt TotalExt
#> 1         1      1        2
#> 2         2      2        4
#> 3         2      3        5