Random extinction

Generates a null model by generating random extinction histories and calculating the mean and standard deviation of the accumulated secondary extinctions developed by making n random extinction histories.

RandomExtinctions(
  Network,
  nsim = 10,
  Record = FALSE,
  plot = FALSE,
  SimNum = NULL,
  NetworkType = "Trophic",
  clust.method = "cluster_infomap",
  parallel = FALSE,
  ncores,
  IS = 0,
  Rewiring = FALSE,
  RewiringDist = NULL,
  RewiringProb = 0.5,
  verbose = TRUE
)

Arguments

Network

a network representation as a an adjacency matrix, edgelist, or a network object

nsim

numeric, number of simulations

Record

logical, if TRUE, records every simulation and you can read the raw results in the object FullSims

plot

logical if TRUE, will add a graph to the results

SimNum

numeric, how many nodes to register for primary extinction. By default sets all of them.

NetworkType

a character with the options Trophic and Mutualistic - is used to calculate secondary extinctions.

clust.method

a character with the options cluster_edge_betweenness, cluster_spinglass, cluster_label_prop or cluster_infomap, defaults to cluster_infomap

parallel

if TRUE, it will use parallel procesing, if FALSE (default) it will run sequentially

ncores

numeric, number of cores to use if using parallel procesing

IS

either numeric or a named vector of numerics. Identifies the threshold of relative interaction strength which species require to not be considered secondarily extinct (i.e. IS = 0.3 leads to removal of all nodes which lose 70 precent of their interaction strength in the Network argument). If a named vector, names must correspond to vertex names in Network argument.

Rewiring

either a function or a named vector of functions. Signifies how rewiring probabilities are calculated from the RewiringDist argument. If FALSE, no rewiring is carried out.

RewiringDist

a numeric matrix of NxN dimension (N... number of nodes in Network). Contains, for example, phylogenetic or functional trait distances between nodes in Network which are used by the Rewiring argument to calculate rewiring probabilities.

RewiringProb

a numeric which identifies the threshold at which to assume rewiring potential is met.

verbose

Logical. Whether to report on function progress or not.

Value

exports list containing a data frame with the characteristics of the network after every extinction, a network object containing the final network, and a graph with the mean and 95percent interval. The resulting data frame contains 11 columns that incorporate the topological index, the secondary extinctions, predation release, and total extinctions of the network in each primary extinction.

Details

When NetworkType = Trophic, secondary extinctions only occur for any predator, but not producers. If NetworkType = Mutualistic, secondary extinctions occur for all species in the network.

When clust.method = cluster_edge_betweenness computes the network modularity using cluster_edge_betweenness methods from igraph to detect communities When clust.method = cluster_spinglass computes the network modularity using cluster_spinglass methods from igraph to detect communities, here the number of spins are equal to the network size When clust.method = cluster_label_prop computes the network modularity using cluster_label_prop methods from igraph to detect communities When clust.method = cluster_infomap computes the network modularity using cluster_infomap methods from igraph to detect communities, here the number of nb.trials are equal to the network size

Author

Derek Corcoran <derek.corcoran.barrios@gmail.com>

M. Isidora Ávila-Thieme <msavila@uc.cl>

Erik Kusch <erik.kusch@bio.au.dk>

Examples

#first example
data("More_Connected")
RandomExtinctions(Network = More_Connected, nsim = 20)
#>   |                                                                              |                                                                      |   0%
#> Warning: Density is not well-defined for networks of order 0.
#> Warning: Density is not well-defined for networks of order 0.
#> Warning: Your network become completely unconnected before all primary extinctions were simulated. This happened at extinction step 27 out of 30
#>   |                                                                              |====                                                                  |   5%
#> Warning: Density is not well-defined for networks of order 0.
#> Warning: Your network become completely unconnected before all primary extinctions were simulated. This happened at extinction step 27 out of 30
#>   |                                                                              |=======                                                               |  10%
#> Warning: Density is not well-defined for networks of order 0.
#> Warning: Density is not well-defined for networks of order 0.
#> Warning: Your network become completely unconnected before all primary extinctions were simulated. This happened at extinction step 29 out of 30
#>   |                                                                              |==========                                                            |  15%
#> Warning: Your network become completely unconnected before all primary extinctions were simulated. This happened at extinction step 26 out of 30
#>   |                                                                              |==============                                                        |  20%
#> Warning: Density is not well-defined for networks of order 0.
#> Warning: Your network become completely unconnected before all primary extinctions were simulated. This happened at extinction step 24 out of 30
#>   |                                                                              |==================                                                    |  25%
#> Warning: Density is not well-defined for networks of order 0.
#> Warning: Density is not well-defined for networks of order 0.
#> Warning: Your network become completely unconnected before all primary extinctions were simulated. This happened at extinction step 25 out of 30
#>   |                                                                              |=====================                                                 |  30%
#> Warning: Density is not well-defined for networks of order 0.
#> Warning: Your network become completely unconnected before all primary extinctions were simulated. This happened at extinction step 27 out of 30
#>   |                                                                              |========================                                              |  35%
#> Warning: Density is not well-defined for networks of order 0.
#> Warning: Density is not well-defined for networks of order 0.
#> Warning: Your network become completely unconnected before all primary extinctions were simulated. This happened at extinction step 28 out of 30
#>   |                                                                              |============================                                          |  40%
#> Warning: Your network become completely unconnected before all primary extinctions were simulated. This happened at extinction step 25 out of 30
#>   |                                                                              |================================                                      |  45%
#> Warning: Density is not well-defined for networks of order 0.
#>   |                                                                              |===================================                                   |  50%
#> Warning: Density is not well-defined for networks of order 0.
#> Warning: Density is not well-defined for networks of order 0.
#> Warning: Your network become completely unconnected before all primary extinctions were simulated. This happened at extinction step 27 out of 30
#>   |                                                                              |======================================                                |  55%
#> Warning: Density is not well-defined for networks of order 0.
#> Warning: Density is not well-defined for networks of order 0.
#> Warning: Your network become completely unconnected before all primary extinctions were simulated. This happened at extinction step 27 out of 30
#>   |                                                                              |==========================================                            |  60%
#> Warning: Density is not well-defined for networks of order 0.
#> Warning: Your network become completely unconnected before all primary extinctions were simulated. This happened at extinction step 27 out of 30
#>   |                                                                              |==============================================                        |  65%
#> Warning: Density is not well-defined for networks of order 0.
#>   |                                                                              |=================================================                     |  70%
#> Warning: Density is not well-defined for networks of order 0.
#> Warning: Density is not well-defined for networks of order 0.
#> Warning: Your network become completely unconnected before all primary extinctions were simulated. This happened at extinction step 23 out of 30
#>   |                                                                              |====================================================                  |  75%
#> Warning: Density is not well-defined for networks of order 0.
#> Warning: Your network become completely unconnected before all primary extinctions were simulated. This happened at extinction step 23 out of 30
#>   |                                                                              |========================================================              |  80%
#> Warning: Your network become completely unconnected before all primary extinctions were simulated. This happened at extinction step 27 out of 30
#>   |                                                                              |============================================================          |  85%
#> Warning: Density is not well-defined for networks of order 0.
#> Warning: Density is not well-defined for networks of order 0.
#> Warning: Your network become completely unconnected before all primary extinctions were simulated. This happened at extinction step 27 out of 30
#>   |                                                                              |===============================================================       |  90%
#> Warning: Density is not well-defined for networks of order 0.
#>   |                                                                              |==================================================================    |  95%
#> Warning: Density is not well-defined for networks of order 0.
#> Warning: Your network become completely unconnected before all primary extinctions were simulated. This happened at extinction step 26 out of 30
#>   |                                                                              |======================================================================| 100%
#> $sims
#> # A tibble: 30 × 5
#>    NumExt AccSecExt_95CI AccSecExt_mean Upper Lower
#>     <int>          <dbl>          <dbl> <dbl> <dbl>
#>  1      1          0               0    0         0
#>  2      2          0.603           0.1  0.703     0
#>  3      3          1.35            0.5  1.85      0
#>  4      4          1.94            0.85 2.79      0
#>  5      5          2.40            1.15 3.55      0
#>  6      6          3.27            1.5  4.77      0
#>  7      7          3.95            1.8  5.75      0
#>  8      8          4.49            2.25 6.74      0
#>  9      9          4.56            2.45 7.01      0
#> 10     10          4.60            2.65 7.25      0
#> # … with 20 more rows
#> 
#> $nets
#> $nets[[1]]
#>  Network attributes:
#>   vertices = 0 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = TRUE 
#>   multiple = FALSE 
#>   bipartite = FALSE 
#>   total edges= 0 
#>     missing edges= 0 
#>     non-missing edges= 0 
#> 
#> No vertex attributes
#> 
#> No edge attributes
#> 
#> $nets[[2]]
#>  Network attributes:
#>   vertices = 0 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = TRUE 
#>   multiple = FALSE 
#>   bipartite = FALSE 
#>   total edges= 0 
#>     missing edges= 0 
#>     non-missing edges= 0 
#> 
#> No vertex attributes
#> 
#> No edge attributes
#> 
#> $nets[[3]]
#>  Network attributes:
#>   vertices = 0 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = TRUE 
#>   multiple = FALSE 
#>   bipartite = FALSE 
#>   total edges= 0 
#>     missing edges= 0 
#>     non-missing edges= 0 
#> 
#> No vertex attributes
#> 
#> No edge attributes
#> 
#> $nets[[4]]
#>  Network attributes:
#>   vertices = 1 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = TRUE 
#>   multiple = FALSE 
#>   bipartite = FALSE 
#>   total edges= 0 
#>     missing edges= 0 
#>     non-missing edges= 0 
#> 
#>  Vertex attribute names: 
#>     vertex.names 
#> 
#> No edge attributes
#> 
#> $nets[[5]]
#>  Network attributes:
#>   vertices = 0 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = TRUE 
#>   multiple = FALSE 
#>   bipartite = FALSE 
#>   total edges= 0 
#>     missing edges= 0 
#>     non-missing edges= 0 
#> 
#> No vertex attributes
#> 
#> No edge attributes
#> 
#> $nets[[6]]
#>  Network attributes:
#>   vertices = 0 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = TRUE 
#>   multiple = FALSE 
#>   bipartite = FALSE 
#>   total edges= 0 
#>     missing edges= 0 
#>     non-missing edges= 0 
#> 
#> No vertex attributes
#> 
#> No edge attributes
#> 
#> $nets[[7]]
#>  Network attributes:
#>   vertices = 0 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = TRUE 
#>   multiple = FALSE 
#>   bipartite = FALSE 
#>   total edges= 0 
#>     missing edges= 0 
#>     non-missing edges= 0 
#> 
#> No vertex attributes
#> 
#> No edge attributes
#> 
#> $nets[[8]]
#>  Network attributes:
#>   vertices = 0 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = TRUE 
#>   multiple = FALSE 
#>   bipartite = FALSE 
#>   total edges= 0 
#>     missing edges= 0 
#>     non-missing edges= 0 
#> 
#> No vertex attributes
#> 
#> No edge attributes
#> 
#> $nets[[9]]
#>  Network attributes:
#>   vertices = 1 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = TRUE 
#>   multiple = FALSE 
#>   bipartite = FALSE 
#>   total edges= 0 
#>     missing edges= 0 
#>     non-missing edges= 0 
#> 
#>  Vertex attribute names: 
#>     vertex.names 
#> 
#> No edge attributes
#> 
#> $nets[[10]]
#>  Network attributes:
#>   vertices = 0 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = TRUE 
#>   multiple = FALSE 
#>   bipartite = FALSE 
#>   total edges= 0 
#>     missing edges= 0 
#>     non-missing edges= 0 
#> 
#> No vertex attributes
#> 
#> No edge attributes
#> 
#> $nets[[11]]
#>  Network attributes:
#>   vertices = 0 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = TRUE 
#>   multiple = FALSE 
#>   bipartite = FALSE 
#>   total edges= 0 
#>     missing edges= 0 
#>     non-missing edges= 0 
#> 
#> No vertex attributes
#> 
#> No edge attributes
#> 
#> $nets[[12]]
#>  Network attributes:
#>   vertices = 0 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = TRUE 
#>   multiple = FALSE 
#>   bipartite = FALSE 
#>   total edges= 0 
#>     missing edges= 0 
#>     non-missing edges= 0 
#> 
#> No vertex attributes
#> 
#> No edge attributes
#> 
#> $nets[[13]]
#>  Network attributes:
#>   vertices = 0 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = TRUE 
#>   multiple = FALSE 
#>   bipartite = FALSE 
#>   total edges= 0 
#>     missing edges= 0 
#>     non-missing edges= 0 
#> 
#> No vertex attributes
#> 
#> No edge attributes
#> 
#> $nets[[14]]
#>  Network attributes:
#>   vertices = 0 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = TRUE 
#>   multiple = FALSE 
#>   bipartite = FALSE 
#>   total edges= 0 
#>     missing edges= 0 
#>     non-missing edges= 0 
#> 
#> No vertex attributes
#> 
#> No edge attributes
#> 
#> $nets[[15]]
#>  Network attributes:
#>   vertices = 0 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = TRUE 
#>   multiple = FALSE 
#>   bipartite = FALSE 
#>   total edges= 0 
#>     missing edges= 0 
#>     non-missing edges= 0 
#> 
#> No vertex attributes
#> 
#> No edge attributes
#> 
#> $nets[[16]]
#>  Network attributes:
#>   vertices = 0 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = TRUE 
#>   multiple = FALSE 
#>   bipartite = FALSE 
#>   total edges= 0 
#>     missing edges= 0 
#>     non-missing edges= 0 
#> 
#> No vertex attributes
#> 
#> No edge attributes
#> 
#> $nets[[17]]
#>  Network attributes:
#>   vertices = 1 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = TRUE 
#>   multiple = FALSE 
#>   bipartite = FALSE 
#>   total edges= 0 
#>     missing edges= 0 
#>     non-missing edges= 0 
#> 
#>  Vertex attribute names: 
#>     vertex.names 
#> 
#> No edge attributes
#> 
#> $nets[[18]]
#>  Network attributes:
#>   vertices = 0 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = TRUE 
#>   multiple = FALSE 
#>   bipartite = FALSE 
#>   total edges= 0 
#>     missing edges= 0 
#>     non-missing edges= 0 
#> 
#> No vertex attributes
#> 
#> No edge attributes
#> 
#> $nets[[19]]
#>  Network attributes:
#>   vertices = 0 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = TRUE 
#>   multiple = FALSE 
#>   bipartite = FALSE 
#>   total edges= 0 
#>     missing edges= 0 
#>     non-missing edges= 0 
#> 
#> No vertex attributes
#> 
#> No edge attributes
#> 
#> $nets[[20]]
#>  Network attributes:
#>   vertices = 0 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = TRUE 
#>   multiple = FALSE 
#>   bipartite = FALSE 
#>   total edges= 0 
#>     missing edges= 0 
#>     non-missing edges= 0 
#> 
#> No vertex attributes
#> 
#> No edge attributes
#> 
#> 

# Using parallel procesing
## Detect your number of cores divide by 2
if (FALSE) {
cores <- ceiling(parallel::detectCores()/2)

RandomExtinctions(Network = More_Connected, nsim = 20, parallel = TRUE, ncores = cores)
}